controllers

+++ src/repo/holiday/index.js
@@ -1,3 +1,4 @@
+const { isValidObjectId } = require("mongoose");
 const { Holiday } = require("../../models");
 const quryReplacer = require("../../utils/quryReplacer");
 
@@ -51,6 +52,14 @@ const findItemById = async ({ id, select = "" }) => {
 };
 
 const updateItem = async ({ qry = {}, updateDate = {}, options = {} }) => {
+  if (
+    !updateDate ||
+    !Object.keys(updateDate).length ||
+    !qry ||
+    !Object.keys(qry).length
+  ) {
+    throw new Error(`Please provide data`);
+  }
   qry = quryReplacer(qry);
   const updated = await Holiday.updateOne(qry, updateDate, options);
   if (!updated.matchedCount) return false;
@@ -58,6 +67,11 @@ const updateItem = async ({ qry = {}, updateDate = {}, options = {} }) => {
 };
 
 const updateItemById = async ({ id, updateDate = {}, options = {} }) => {
+  if (!updateDate || !Object.keys(updateDate).length) {
+    throw new Error(`Please provide data`);
+  } else if (!id || !isValidObjectId(id)) {
+    throw new Error(`Please provide valid ID!`);
+  }
   const updated = await Holiday.findByIdAndUpdate(id, updateDate, options);
   if (!updated) return false;
   const copy = { id: updated.id, ...updated._doc };
@@ -66,36 +80,50 @@ const updateItemById = async ({ id, updateDate = {}, options = {} }) => {
 };
 
 const deleteItem = ({ qry = {} }) => {
+  if (!qry || !Object.keys(qry).length) {
+    throw new Error(`Please provide data`);
+  }
   qry = quryReplacer(qry);
   return Holiday.deleteOne(qry);
 };
 
 const deleteItemById = ({ id }) => {
+  if (!id || !isValidObjectId(id)) {
+    throw new Error(`Please provide valid ID!`);
+  }
   return Holiday.findByIdAndDelete(id);
 };
 
 const deleteManyItem = ({ qry = {} }) => {
+  if (!data || !Object.keys(qry).length) {
+    throw new Error(`Please provide data`);
+  }
   qry = quryReplacer(qry);
   return Holiday.deleteMany(qry);
 };
 
 const createNewItem = async ({ data }) => {
+  if (!data || !Object.keys(data).length) {
+    throw new Error(`Please provide data`);
+  }
   const newData = new Holiday(data);
   await newData.save();
-  const copy = newData._doc;
+  const copy = JSON.parse(JSON.stringify(newData._doc));
   delete copy._id;
   return { id: newData.id, ...copy };
 };
 
 const create = async ({ ...data }) => {
-  const updated = await Holiday.create(data);
-  const copy = { id: updated.id, ...updated._doc };
+  if (!data || !Object.keys(data).length) {
+    throw new Error(`Please provide data`);
+  }
+  const created = await Holiday.create(data);
+  const copy = { id: created.id, ...created._doc };
   delete copy._id;
   delete copy.password;
   delete copy.__v;
   return copy;
 };
-
 const count = ({ filter = {} }) => {
   filter = quryReplacer(filter);
   return Holiday.countDocuments(filter);
